+++ include/falcon/core/FalconComm.h
@@ -7,8 +7,8 @@
  * @license BSD License
  *
  * $HeadURL$
- * 
- * Project info at http://libnifalcon.sourceforge.net/ 
+ *
+ * Project info at http://libnifalcon.sourceforge.net/
  *
  */
 #ifndef FALCONCOMMBASE_H
@@ -17,10 +17,10 @@
 #include <stdint.h>
 #include "falcon/core/FalconCore.h"
 
-namespace libnifalcon 
+namespace libnifalcon
 {
 	class FalconComm : public FalconCore
-	{		
+	{
 	public:
 		enum {
 			FALCON_COMM_DEVICE_ERROR = 2000,  /*!< Device driver specific error, check getDeviceErrorCode */
@@ -33,10 +33,10 @@ namespace libnifalcon
 			FALCON_COMM_READ_ERROR /*!< Read timeout hit, underflow, etc... */
 		};
 
-		/** 
+		/**
 		 * Constructor
-		 * 
-		 * 
+		 *
+		 *
 		 */
 		FalconComm() :
 			m_isCommOpen(false),
@@ -44,93 +44,93 @@ namespace libnifalcon
 			m_hasBytesAvailable(false),
 			m_bytesAvailable(0)
 		{}
-		/** 
+		/**
 		 * Destructor
-		 * 
-		 * 
+		 *
+		 *
 		 */
 		virtual ~FalconComm() {}
-		/** 
+		/**
 		 * Returns the number of devices connected to the system
-		 * 
+		 *
 		 * @param[out] count The number of devices available
-		 * 
+		 *
 		 * @return True if count was retreived correctly, false otherwise. Error code set if false.
 		 */
-		virtual bool getDeviceCount(int8_t& count) = 0;
-		/** 
+		virtual bool getDeviceCount(unsigned int& count) = 0;
+		/**
 		 * Opens the device at the specified index
-		 * 
+		 *
 		 * @param[in] index Index of the device to open
-		 * 
+		 *
 		 * @return True if device is opened successfully, false otherwise. Error code set if false.
 		 */
-		virtual bool open(uint8_t index) = 0;
-		/** 
+		virtual bool open(unsigned int index) = 0;
+		/**
 		 * Closes the device, if open
-		 * 
-		 * 
+		 *
+		 *
 		 * @return True if device is closed successfully, false otherwise. Error code set if false.
 		 */
 		virtual bool close() = 0;
-		/** 
+		/**
 		 * Read a specified number of bytes from the device
-		 * 
+		 *
 		 * @param[out] str Buffer to read data into
 		 * @param[in] size Amount of bytes to read
-		 * 
+		 *
 		 * @return True if (size) amount of bytes is read successfully, false otherwise. Error code set if false.
 		 */
-		virtual bool read(uint8_t* str, uint32_t size) = 0;
-		/** 
+		virtual bool read(uint8_t* str, unsigned int size) = 0;
+		/**
 		 * Write a specified number of bytes to the device
-		 * 
+		 *
 		 * @param[in] str Buffer to write data from
 		 * @param[in] size Amount of bytes to write
-		 * 
+		 *
 		 * @return True if (size) amount of bytes is written successfully, false otherwise. Error code set if false.
 		 */
-		virtual bool write(uint8_t* str, uint32_t size) = 0;
-		/** 
-		 * Sets the communications mode and initializes the device to load firmware 
-		 * 
-		 * 
+		virtual bool write(uint8_t* str, unsigned int size) = 0;
+		/**
+		 * Sets the communications mode and initializes the device to load firmware
+		 *
+		 *
 		 * @return True if device is successfully set to load firwmare, false otherwise. Error code set if false.
 		 */
 		virtual bool setFirmwareMode() = 0;
-		/** 
+		/**
 		 * Sets the communications mode and initializes the device to run in normal operation
-		 * 
-		 * 
+		 *
+		 *
 		 * @return True if device is successfully set to normal operation, false otherwise. Error code set if false.
-		 */		
+		 */
 		virtual bool setNormalMode() = 0;
 
-		/** 
+		/**
 		 * Returns the number of bytes returned in the last read function
-		 * 
-		 * 
+		 *
+		 *
 		 * @return Number of bytes read
 		 */
 		int getLastBytesRead()  {return m_lastBytesRead; }
-		/** 
+		/**
 		 * Returns the number of bytes returned in the last write function
-		 * 
-		 * 
+		 *
+		 *
 		 * @return Number of bytes written
 		 */
 		int getLastBytesWritten()  {return m_lastBytesWritten; }
-		/** 
+		/**
 		 * Returns the device specific error code for in-depth debugging
-		 * 
-		 * 
+		 *
+		 *
 		 * @return Index for the communication policy specific error code
 		 */
 		int getDeviceErrorCode() { return m_deviceErrorCode; }
-		/** 
+		/**
 		 * Returns whether the device is open or not
-		 * 
-		 * 
+		 *
+		 *
 		 * @return True if open, false otherwise
 		 */
 		bool isCommOpen() { return m_isCommOpen; }
